+++ config.go
@@ -40,7 +40,7 @@ type Config struct {
 	StartFromLSN uint64 `envconfig:"START_FROM_LSN"`
 
 	// Start replication after the changeset count offset. (Audit mode only)
-	StartFromOffset int `envconfig:"START_FROM_OFFSET"`
+	StartFromOffset int64 `envconfig:"START_FROM_OFFSET"`
 
 	// Start replication from the specified changeset timestamp. (Audit mode only)
 	StartFromTimestamp int64 `envconfig:"START_FROM_TIMESTAMP"`
