+++ internal/cli/warp_pipe.go
@@ -25,7 +25,7 @@ var (
 	replicationMode    string
 	ignoreTables       []string
 	whitelistTables    []string
-	startFromOffset    int
+	startFromOffset    int64
 	startFromTimestamp int64
 	startFromLSN       int64
 	logLevel           string
@@ -44,7 +44,7 @@ func init() {
 	WarpPipeCmd.PersistentFlags().StringVarP(&dbUser, "db-user", "U", "", "database user")
 	WarpPipeCmd.PersistentFlags().StringVarP(&dbPass, "db-pass", "P", "", "database password")
 	WarpPipeCmd.Flags().Int64Var(&startFromLSN, "start-from-lsn", -1, "stream all changes starting from the provided LSN")
-	WarpPipeCmd.Flags().IntVar(&startFromOffset, "start-from-offset", -1, "stream all changes starting from the provided changeset offset")
+	WarpPipeCmd.Flags().Int64Var(&startFromOffset, "start-from-offset", -1, "stream all changes starting from the provided changeset offset")
 	WarpPipeCmd.Flags().Int64Var(&startFromTimestamp, "start-from-ts", -1, "stream all changes starting from the provided timestamp")
 	WarpPipeCmd.Flags().StringVarP(&replicationMode, "replication-mode", "M", replicationModeLR, "replication mode")
 	WarpPipeCmd.Flags().StringSliceVarP(&ignoreTables, "ignore-tables", "i", nil, "tables to ignore during replication")
